requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
requests>=2.28.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
langchain-openai>=0.3.35